feat: show themes status on home
